Creates runner file for RPN class.
